+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* 2x.x.x
+  - added sapyb and deprecated axpby
+
 * 21.3.1
   - Added matplotlib version dependency to conda recipe
   - Fixed TIGRE wrappers for geometry with a virtual detector